backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Redis
redis==5.0.1

# HTTP client
httpx==0.26.0
tenacity==8.2.3

# Data validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Date/time
python-dateutil==2.8.2

# Scheduling
apscheduler==3.10.4

# Export
openpyxl==3.1.2
xlsxwriter==3.1.9

# Logging
structlog==24.1.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov>=4.0.0

# Code quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Utils
python-dotenv==1.0.1
orjson==3.9.13
